moving New Collection Dialog, localizing header
